Refactor test structure to reflect changes
