+++ ethpm/_utils/chains.py
@@ -3,8 +3,7 @@ from typing import Tuple
 from urllib import parse
 
 from eth_typing import URI
-from eth_utils import add_0x_prefix, encode_hex, remove_0x_prefix, to_hex
-from eth_utils.toolz import curry
+from eth_utils import add_0x_prefix, remove_0x_prefix
 from web3 import Web3
