+++ ethpm/backends/http.py
@@ -1,20 +1,16 @@
 import base64
-import hashlib
 import json
 from typing import Tuple
 from urllib import parse
 
 from eth_typing import URI
-from eth_utils import is_text, to_bytes, to_text
+from eth_utils import is_text
 import requests
 
 from ethpm.backends.base import BaseURIBackend
 from ethpm.constants import GITHUB_API_AUTHORITY
-from ethpm.exceptions import CannotHandleURI, ValidationError
-from ethpm.utils.uri import (
-    is_valid_content_addressed_github_uri,
-    validate_blob_uri_contents,
-)
+from ethpm.exceptions import CannotHandleURI
+from ethpm.validation.uri import validate_blob_uri_contents
 
 
 class GithubOverHTTPSBackend(BaseURIBackend):
@@ -94,21 +90,3 @@ def is_valid_github_uri(uri: URI, expected_path_terms: Tuple[str, ...]) -> bool:
     if authority != GITHUB_API_AUTHORITY:
         return False
     return True
-
-
-def validate_blob_uri_contents(contents: bytes, blob_uri: str) -> None:
-    """
-    Raises an exception if the sha1 hash of the contents does not match the hash found in te
-    blob_uri. Formula for how git calculates the hash found here:
-    http://alblue.bandlem.com/2011/08/git-tip-of-week-objects.html
-    """
-    blob_path = parse.urlparse(blob_uri).path
-    blob_hash = blob_path.split("/")[-1]
-    contents_str = to_text(contents)
-    content_length = len(contents_str)
-    hashable_contents = "blob " + str(content_length) + "\0" + contents_str
-    hash_object = hashlib.sha1(to_bytes(text=hashable_contents))
-    if hash_object.hexdigest() != blob_hash:
-        raise ValidationError(
-            f"Hash of contents fetched from {blob_uri} do not match its hash: {blob_hash}."
-        )
