+++ ethpm/package.py
@@ -8,6 +8,7 @@ from web3 import Web3
 from web3.eth import Contract
 
 from ethpm._utils.cache import cached_property
+from ethpm._utils.contract import generate_contract_factory_kwargs
 from ethpm._utils.deployments import (
     get_linked_deployments,
     normalize_linked_references,
@@ -24,9 +25,6 @@ from ethpm.exceptions import (
     PyEthPMError,
 )
 from ethpm.uri import resolve_uri_contents
-from ethpm._utils.contract import (
-    generate_contract_factory_kwargs,
-)
 from ethpm.validation.manifest import (
     check_for_deployments,
     validate_build_dependencies_are_present,
@@ -34,18 +32,13 @@ from ethpm.validation.manifest import (
     validate_manifest_deployments,
     validate_raw_manifest_format,
 )
+from ethpm.validation.misc import validate_address, validate_w3_instance
 from ethpm.validation.package import (
-    validate_minimal_contract_factory_data,
     validate_build_dependency,
     validate_contract_name,
+    validate_minimal_contract_factory_data,
 )
-from ethpm.validation.uri import (
-    validate_single_matching_uri,
-)
-from ethpm.validation.misc import (
-    validate_address,
-    validate_w3_instance,
-)
+from ethpm.validation.uri import validate_single_matching_uri
 
 
 class Package(object):
