+++ tests/ethpm/_utils/test_contract_utils.py
@@ -1,11 +1,12 @@
 import pytest
 
+from ethpm._utils.contract import generate_contract_factory_kwargs
 from ethpm.exceptions import InsufficientAssetsError, ValidationError
-from ethpm._utils.contract import (
-    generate_contract_factory_kwargs,
-)
 from ethpm.validation.misc import validate_w3_instance
-from ethpm.validation.package import validate_contract_name, validate_minimal_contract_factory_data
+from ethpm.validation.package import (
+    validate_contract_name,
+    validate_minimal_contract_factory_data,
+)
 
 
 @pytest.mark.parametrize(
