+++ tests/ethpm/_utils/test_registry_utils.py
