+++ tests/ethpm/test_contract.py
@@ -5,7 +5,6 @@ from web3.contract import Contract
 from ethpm import Package
 from ethpm.contract import LinkableContract, apply_all_link_refs
 from ethpm.exceptions import BytecodeLinkingError, ValidationError
-from ethpm.validation.misc import validate_empty_bytes
 
 
 @pytest.mark.parametrize(
@@ -177,30 +176,3 @@ def test_unlinked_factory_cannot_be_deployed(get_factory):
     assert escrow.needs_bytecode_linking
     with pytest.raises(BytecodeLinkingError):
         escrow.constructor("0x4F5B11c860b37b68DE6D14Fb7e7b5f18A9A1bdC0").transact()
-
-
-@pytest.mark.parametrize(
-    "offset,length,bytecode",
-    (
-        (0, 3, b"\00\00\00"),
-        (1, 20, b"\01" + bytearray(20) + b"\01"),
-        (26, 20, b"\01" + bytearray(20) + b"\01" * 5 + bytearray(20) + b"\01"),
-    ),
-)
-def test_validate_empty_bytes(offset, length, bytecode):
-    result = validate_empty_bytes(offset, length, bytecode)
-    assert result is None
-
-
-@pytest.mark.parametrize(
-    "offset,length,bytecode",
-    (
-        (0, 2, b"\00"),
-        (0, 3, b"\01\01\01"),
-        (1, 1, b"\00\01\00\01"),
-        (1, 20, bytearray(20) + b"\01"),
-    ),
-)
-def test_validate_empty_bytes_invalidates(offset, length, bytecode):
-    with pytest.raises(ValidationError):
-        validate_empty_bytes(offset, length, bytecode)
