+++ tests/ethpm/test_uri.py
