+++ tests/ethpm/validation/test_manifest.py
@@ -7,6 +7,7 @@ from ethpm.validation.manifest import (
     validate_manifest_against_schema,
     validate_manifest_deployments,
     validate_manifest_exists,
+    validate_meta_object,
     validate_raw_manifest_format,
 )
 from ethpm.validation.package import validate_manifest_version
@@ -111,3 +112,100 @@ def test_validate_manifest_version_validates_version_two_string(version):
 def test_validate_manifest_version_invalidates_incorrect_versions(version):
     with pytest.raises(ValidationError):
         validate_manifest_version(version)
+
+
+@pytest.mark.parametrize(
+    "meta,extra_fields",
+    (
+        (
+            {
+                "license": "MIT",
+                "authors": ["author@gmail.com"],
+                "description": "A Package that does things.",
+                "keywords": ["ethpm", "package"],
+                "links": {"documentation": "ipfs://Qm..."},
+            },
+            False,
+        ),
+        (
+            {
+                "license": "MIT",
+                "authors": ["author@gmail.com"],
+                "description": "A Package that does things.",
+                "keywords": ["ethpm", "package"],
+                "links": {"documentation": "ipfs://Qm..."},
+                "x-hash": "0x...",
+            },
+            True,
+        ),
+    ),
+)
+def test_validate_meta_object_validates(meta, extra_fields):
+    result = validate_meta_object(meta, allow_extra_meta_fields=extra_fields)
+    assert result is None
+
+
+@pytest.mark.parametrize(
+    "meta,extra_fields",
+    (
+        # With allow_extra_meta_fields=False
+        ({"invalid": "field"}, False),
+        ({"license": 123}, False),
+        ({"license": "MIT", "authors": "auther@gmail.com"}, False),
+        (
+            {
+                "license": "MIT",
+                "authors": ["author@gmail.com"],
+                "description": ["description", "of", "package"],
+            },
+            False,
+        ),
+        (
+            {
+                "license": "MIT",
+                "authors": ["author@gmail.com"],
+                "description": "description",
+                "keywords": "singlekw",
+            },
+            False,
+        ),
+        (
+            {
+                "license": "MIT",
+                "authors": ["author@gmail.com"],
+                "description": "description",
+                "keywords": ["auth", "package"],
+                "links": ["ipfs://Qm"],
+            },
+            False,
+        ),
+        (
+            {
+                "license": "MIT",
+                "authors": ["author@gmail.com"],
+                "description": "description",
+                "keywords": ["auth", "package"],
+                "links": {"documentation": "ipfs://Qm"},
+                "extra": "field",
+            },
+            False,
+        ),
+        (
+            {
+                "license": "MIT",
+                "authors": ["author@gmail.com"],
+                "description": "description",
+                "keywords": ["auth", "package"],
+                "links": {"documentation": "ipfs://Qm"},
+                "x-hash": "0x",
+            },
+            False,
+        ),
+        # With allow_extra_meta_fields=True
+        # Improperly formatted "x" field
+        ({"license": "MIT", "extra": "field"}, True),
+    ),
+)
+def test_validate_meta_object_invalidates(meta, extra_fields):
+    with pytest.raises(ValidationError):
+        validate_meta_object(meta, allow_extra_meta_fields=extra_fields)
