+++ tests/ethpm/validation/test_manifest_assets_are_valid.py
